Move chain utils
