+++ src/blockchain/block.ts
@@ -1,5 +1,5 @@
 import SHA256 from 'crypto-js/sha256'
-import ChainUtil from '../chain-util'
+import ChainUtil from '../utils/chain-util'
 import Wallet from '../wallet'
 import Transaction from '../wallet/transaction'
